Add change to DayplanViewModel
