+++ src/main/java/view/DayplanView.java
@@ -1,24 +1,25 @@
 package view;
 
+import interface_adapter.DayplanController;
+import interface_adapter.DayplanViewModel;
+
 import java.util.Scanner;
 
 public class DayplanView {
 
-    private String userInput;
+    private final DayplanController dayplanController;
+    private final DayplanViewModel dayplanViewModel;
 
-    /**
-     * Gets String userInput from DayplanView
-     * */
-    public String getUserInput() {
-        return this.userInput;
-    }
 
-    /**
-     * Get userInput from the console
-     * */
-    public void setUserInput() {
+    public DayplanView(DayplanController dayplanController, DayplanViewModel dayplanViewModel) {
+        this.dayplanController = dayplanController;
+        this.dayplanViewModel = dayplanViewModel;
+
+        //add code here that takes whatever userInput and creates a new object out of it
+        String prompt = dayplanViewModel.getPrompt();
         System.out.println(prompt);
-        Scanner input = new Scanner(System.in);
-        this.userInput = input.nextLine();
+        dayplanViewModel.setUserInput();
+        String userInput = dayplanViewModel.getUserInput();
+        dayplanController.execute(userInput);
     }
 }
